+++ watchtower-app/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - 2024-08-19
+### Fixed
+- Display decimal points in chart data - `@bobbles`
+
 ## [1.3.1] - 2024-08-16
 ### Fixed
 - Fix situations where tooltip is obscured by the time resolution picker or the chart title - `@an39511`